requirements.txt
spacy
faiss-cpu
numpy
scikit-learn

